Add test for remove root menu bug
The bug occurs when all rows are removed and the parent actions menu
is removed, too. If the parent actions menu is the menuview itself,
the action and the menu view are never parented back together.

So the next time rows are inserted, it leads to a wierd behavior, where
the menu of the parent action is None.